requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
pandas >2.2.1
django-rq==2.10.2
celery
django-celery-beat
redis
djangorestframework
